+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [Unreleased]
 ### Added
+- adding an ExecutionHandler to host the pointers (instruction, page, stack) and execution related structures (stack, locals, scopes), to ease the transition to a parallelized VM
 
 ### Changed
+- the VM now takes a reference to an `Ark::State` instead of a raw non-owning pointer
 
 ### Removed